refactor: 댓글 DTO 통합 및 단순화
- CreateCommentRequestDto와 CommentRequestDto 통합
- CreateCommentResponseDto와 CommentResponseDto 통합
- 코드 간결화 및 관리 편의성 향상

resolves: #17
