+++ src/main/java/com/sparta/taskflow/domain/comment/controller/CommentController.java
@@ -1,8 +1,7 @@
 package com.sparta.taskflow.domain.comment.controller;
 
-import com.sparta.taskflow.domain.comment.dto.request.CreateCommentRequestDto;
-import com.sparta.taskflow.domain.comment.dto.response.CommentResponseDto;
-import com.sparta.taskflow.domain.comment.dto.response.CreateCommentResponseDto;
+import com.sparta.taskflow.domain.comment.dto.CommentRequestDto;
+import com.sparta.taskflow.domain.comment.dto.CommentResponseDto;
 import com.sparta.taskflow.domain.comment.service.CommentService;
 import com.sparta.taskflow.response.ApiResponse;
 import jakarta.validation.Valid;
@@ -16,42 +15,32 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api")
+@RequestMapping("/api/tasks/{taskId}/comments")
 public class CommentController {
 
     private final CommentService commentService;
 
-    @PostMapping("/tasks/{taskId}/comments")
-    public ResponseEntity<ApiResponse<CreateCommentResponseDto>> createComment(
+    // 댓글 생성
+    @PostMapping
+    public ResponseEntity<ApiResponse<CommentResponseDto>> createComment(
         @PathVariable Long taskId,
-        @Valid @RequestBody CreateCommentRequestDto requestDto
+        @Valid @RequestBody CommentRequestDto requestDto
     ) {
-        CreateCommentResponseDto responseDto = commentService.createComment(requestDto, taskId);
-
-        ApiResponse<CreateCommentResponseDto> response = ApiResponse.success(
-            "댓글이 생성되었습니다.",
-            responseDto
-        );
-
+        CommentResponseDto responseDto = commentService.createComment(requestDto, taskId);
+        ApiResponse<CommentResponseDto> response = ApiResponse.success("댓글이 생성되었습니다.", responseDto);
         return ResponseEntity.status(201).body(response);
     }
 
-    // 댓글 목록 조회 API (페이지네이션 적용, 명세서대로 /api/tasks/{taskId}/comments)
-    @GetMapping("/tasks/{taskId}/comments")
+    // 댓글 목록 조회
+    @GetMapping
     public ResponseEntity<ApiResponse<Page<CommentResponseDto>>> getCommentsByTask(
         @PathVariable Long taskId,
         @RequestParam(defaultValue = "0") int page,
         @RequestParam(defaultValue = "10") int size
     ) {
         Pageable pageable = PageRequest.of(page, size, Sort.by("createdAt").descending());
-        Page<CommentResponseDto> responseDtoPage = commentService.getCommentsByTask(taskId,
-            pageable);
-
-        ApiResponse<Page<CommentResponseDto>> response = ApiResponse.success(
-            "댓글 목록을 조회했습니다.",
-            responseDtoPage
-        );
-
+        Page<CommentResponseDto> responsePage = commentService.getCommentsByTask(taskId, pageable);
+        ApiResponse<Page<CommentResponseDto>> response = ApiResponse.success("댓글 목록을 조회했습니다.", responsePage);
         return ResponseEntity.ok(response);
     }
 }
