+++ src/main/java/com/sparta/taskflow/domain/comment/dto/CommentResponseDto.java
@@ -1,10 +1,11 @@
-package com.sparta.taskflow.domain.comment.dto.response;
+package com.sparta.taskflow.domain.comment.dto;
 
 import com.sparta.taskflow.domain.comment.entity.Comment;
 import com.sparta.taskflow.domain.user.dto.response.UserResponseDto;
-import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
+
+import java.time.LocalDateTime;
 
 @Getter
 @Builder
@@ -14,17 +15,17 @@ public class CommentResponseDto {
     private String content;
     private Long taskId;
     private Long userId;
-    private UserResponseDto user;
+    private UserResponseDto user;  // 사용자 정보 포함
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    public static CommentResponseDto of(Comment comment, UserResponseDto user) {
+    public static CommentResponseDto of(Comment comment, UserResponseDto userDto) {
         return CommentResponseDto.builder()
                                  .id(comment.getId())
                                  .content(comment.getContent())
                                  .taskId(comment.getTaskId())
                                  .userId(comment.getUserId())
-                                 .user(user)
+                                 .user(userDto)
                                  .createdAt(comment.getCreatedAt())
                                  .updatedAt(comment.getUpdatedAt())
                                  .build();
